REFACTOR: Remover arquivos desnecessários numpy_metrics.py e normalized_metrics.py
